chore(notification): Refactor device token logic files
